etl/requirements.txt
backoff==2.1.2
psycopg2-binary==2.9.3
pydantic==1.9.1
python-dotenv==0.20.0
requests==2.28.1